backend/requirements.txt
flask
sympy
flask_cors

alpaca-py
python-dotenv

langgraph
langchain
langchain-ollama

yfinance